Cache index.html for commonplace projects for a few minutes (bug 1071468)
Those pages are served by us, not the CDN, so we can afford to cache
them for a few minutes to help repeat views. If a client wants to bypass
the cache (for instance when we push to production) they can hit reload,
their browser will send a If-Modified-Since or If-None-Match and we'll
give them the new content.