Re-organize trending and popularity tasks
This moves everything from tasks to cron. The tasks were hitting a
recursion error b/c we were chaining too many tasks together.

This also combines the calculating of values and reindexing on the same
chunks of apps and adds a records purge so the values in ES are synced
for purged records.