Store manifest when upload a langpack and simpliy model (bug 1122273)
We need to keep some keys in the minifest identical to the manifest,
before the platform compares the two and will complain if some keys
like name, developer, locales are different.

This means we can drop the custom name property. In addition, some
refactoring was needed to be able to share webapps manifest parsing
and minifest generation, and as a result we're able to drop a few
fields.