Replace all our tasks by post_request_task now that it's a separate lib (bug 1241071)
post-request-task was extracted to zamboni and published in a separate lib. One
important change from before is that it only queues tasks when a request has been
started, which means there is nothing special to do anymore to launch tasks from
a management command or a django shell.

Thanks to this change I modified all our tasks to be post_requests_tasks as there
is no drawback and the added benefit is huge.

Some tests had to be updated to either delay their indexing tasks or reindex after
they did some modifications that aren't automatically triggering reindexes. We used
to not see those because the indexing was delayed or not happening at all in some
cases.